remove duplicated code
